requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]>=2.0.37
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
